Find gears and sum ratios
Easy to double-count gears in this one. The problem with testing the
Line object is it doesn't test the main function, which is where the bug
was. That is, it was wrong to automatically compare l1 to itself when
checking for gears, because when you do that to l2 in main, you do it
again when that becomes l1.

Moral of the story is KISS: compare two lines, and if you need to
compare a line to itself as well, do it separately.